requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
amqpcli==0.0.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.2
cryptography==41.0.3
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
docstring-to-markdown==0.12
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.104.1
fastjsonschema==2.16.3
fqdn==1.5.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.4
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.0
openai==1.3.5
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.0
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.0
pluggy==1.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.14.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.4
pytz==2023.3.post1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
selenium==4.15.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.27.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uri-template==1.2.0
urllib3==2.0.5
uvicorn==0.24.0.post1
vine==5.0.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.5
wsproto==1.2.0
yarl==1.9.2
